+++ src/main/kotlin/com/networkedassets/gherkin/runner/runners/FeatureRunner.kt
@@ -27,7 +27,6 @@ class FeatureRunner(private val implementationsPackage: String,
         featureReport.start()
         try {
             val featureSpecification = Reflection.getFeatureSpecification(implementationsPackage, feature)
-
             runImplemented(featureSpecification)
         } catch (e: NotFoundImplementationException) {
             runNotImplemented()
@@ -39,8 +38,6 @@ class FeatureRunner(private val implementationsPackage: String,
     }
 
     private fun runImplemented(featureSpecification: FeatureSpecification) {
-
-
         beforeFeatureRunner().run(featureSpecification)
         scenarioRunners().forEach { it.run(featureSpecification) }
         afterFeatureRunner().run(featureSpecification)
